fix: move LoggingLevel to own file
